Fix bug that return long instead of int

+++ src/main/java/ua/epam/akoreshev/finalproject/model/dao/CategoryDao.java
@@ -6,7 +6,7 @@ import ua.epam.akoreshev.finalproject.model.entity.Category;
 import java.util.List;
 
 public interface CategoryDao extends BaseDao<Category, Long> {
-    long getNumberCategories() throws DaoException;
+    int getNumberCategories() throws DaoException;
 
     List<Category> findAll(int limit, int offset, String columnName, String sortOrder) throws DaoException;
 }
